requirements.txt
asgiref==3.2.6
autopep8==1.5.3
Babel==2.8.0
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
phonenumberslite==8.12.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
python-dotenv==0.13.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
